+++ src/backend/src/main/java/com/example/backend/data/here/HereApiGeocodeResponse.java
@@ -1,4 +1,4 @@
-package com.example.backend.data.api;
+package com.example.backend.data.here;
 
 import com.example.backend.data.here.Item;
 import com.google.gson.annotations.SerializedName;
