+++ src/backend/src/main/java/com/example/backend/data/osm/OSMQuery.java
@@ -1,4 +1,4 @@
-package com.example.backend.data.api;
+package com.example.backend.data.osm;
 
 import com.example.backend.data.nlp.Height;
