+++ src/backend/src/main/java/com/example/backend/helpers/EmptyResultException.java
@@ -1,0 +1,7 @@
+package com.example.backend.helpers;
+
+public class EmptyResultException extends Throwable {
+    public EmptyResultException(String msg) {
+        super(msg);
+    }
+}
