+++ src/backend/src/main/java/com/example/backend/helpers/Error.java
@@ -1,4 +1,4 @@
-package com.example.backend.data.http;
+package com.example.backend.helpers;
 
 public class Error {
     private final String type;
