+++ src/backend/src/main/java/com/example/backend/helpers/Version.java
@@ -1,4 +1,4 @@
-package com.example.backend.data.http;
+package com.example.backend.helpers;
 
 public class Version {
     private final String backend;
